backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
alembic==1.13.1
pydantic[email]==2.6.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
slowapi==0.1.9
stripe==8.0.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
openai>=1.0.0
